requirements.txt
langchain>=0.1.0
langchain-community>=0.0.30
langchain-openai>=0.0.6
openai>=1.0.0
chromadb>=0.4.22
python-dotenv>=1.0.1
tiktoken>=0.5.1
uuid
langchain_core
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu
streamlit